refactor: 업로드 준비 단계 분리 및 컨텍스트 기반 처리로 정리 - UploadPreparationResult 추가로 파싱/조회 결과 캡슐화 - processRetailExcel에서 ctx 기반 처리 및 스킵 목록 초기 병합
