+++ src/main/java/com/almang/inventory/retail/dto/upload/SkippedRow.java
@@ -1,4 +1,4 @@
-package com.almang.inventory.retail.dto.excel;
+package com.almang.inventory.retail.dto.upload;
 
 public record SkippedRow(
         int rowIndex,
